+++ java/envio/pom.xml
@@ -3,10 +3,11 @@
          xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>com.example</groupId>
-    <artifactId>swing-http-client</artifactId>
+    <groupId>cl.prezdev</groupId>
+    <artifactId>envio</artifactId>
     <version>1.0.0</version>
     <packaging>jar</packaging>
+    <name>Envio</name>
 
     <properties>
         <maven.compiler.release>17</maven.compiler.release>
@@ -43,7 +44,7 @@
                 <configuration>
                     <archive>
                         <manifest>
-                            <mainClass>com.example.httpclient.App</mainClass>
+                            <mainClass>cl.prezdev.envio.App</mainClass>
                         </manifest>
                     </archive>
                 </configuration>
