+++ java/envio/src/main/java/cl/prezdev/envio/App.java
@@ -1,4 +1,4 @@
-package com.example.httpclient;
+package cl.prezdev.envio;
 
 import com.formdev.flatlaf.FlatDarculaLaf;
