+++ java/envio/src/main/java/cl/prezdev/envio/HttpStatusMessageResolver.java
@@ -1,4 +1,4 @@
-package com.example.httpclient;
+package cl.prezdev.envio;
 
 import java.util.Map;
